+++ packages/app/hivecommand-frontend/src/pages/device-control/utils/reports.ts
@@ -19,8 +19,12 @@ export const useDeviceReports = (id: string) => {
 
           recurring
 
+          createdAt 
+
           fields {
             id
+
+            createdAt
 
             device {
               id