+++ packages/editors/interface/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.299](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.298...v1.4.15-alpha.299) (2024-03-14)
+
+**Note:** Version bump only for package @hive-command/interface-editor
+
+
+
+
+
+## [1.4.15-alpha.298](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.297...v1.4.15-alpha.298) (2024-03-14)
+
+**Note:** Version bump only for package @hive-command/interface-editor
+
+
+
+
+
 ## [1.4.15-alpha.297](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.296...v1.4.15-alpha.297) (2024-03-14)
 
 **Note:** Version bump only for package @hive-command/interface-editor